Set GetScenarioTree function into one utils file
